refactor: default torcontrol port to uint16_t

+++ src/torcontrol.h
@@ -19,7 +19,7 @@
 #include <string>
 #include <vector>
 
-constexpr int DEFAULT_TOR_CONTROL_PORT = 9051;
+constexpr uint16_t DEFAULT_TOR_CONTROL_PORT = 9051;
 extern const std::string DEFAULT_TOR_CONTROL;
 static const bool DEFAULT_LISTEN_ONION = true;
